Split out schema_statement and postgresql_adapter patches into separate files.
The require(s) are necessary since Rails doesn’t load connection adapter or db adapter related code until it attempts to establish a DB connection. We know that we’re using PostgreSQL downstream so requiring these doesn’t feel “dirty”, but prudent.
